Create tests to validate actual cnxml
This tests the validation of cnxml using actual cnxml.
