--- cnxml/tests/test_validate.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-from pathlib import Path
-
-
-here = Path(__file__).parent
-DATA_DIR = here / 'data'
-
-
-def test_validate_cnxml():
-    from cnxml import validate_cnxml
-    errors = validate_cnxml(DATA_DIR / 'valid.cnxml')
-    assert errors == tuple()
